wip: restructure
